Suppor table headings and fill columns.
Headings are displayed in bold in a separate row; when the
size of a column is set to -1.0, all the space not used by
other columns is given to that one.

There can be only one fill column in a table, of course;
there are currently no checks to make sure this condition
is satisfied.